+++ modules/gcp/project-factory/variables.tf
@@ -127,9 +127,21 @@ variable "github_repository" {
   type        = string
   default     = null
   validation {
-    condition = var.enable_wif == false || (var.github_repository != null && can(regex("^[a-zA-Z0-9._-]+/[a-zA-Z0-9._-]+$", var.github_repository)))
+    condition     = var.enable_wif == false || (var.github_repository != null && can(regex("^[a-zA-Z0-9._-]+/[a-zA-Z0-9._-]+$", var.github_repository)))
     error_message = "GitHub repository must be in 'owner/repo' format when enable_wif is true."
   }
+}
+
+variable "apply_branch_pattern" {
+  description = "The git branch pattern that is allowed to impersonate the Applier SA. E.g., 'refs/heads/main'."
+  type        = string
+  default     = "refs/heads/main"
+}
+
+variable "plan_branch_pattern" {
+  description = "The git branch pattern that is allowed to impersonate the Planner SA. E.g., 'refs/pull/*' for PRs."
+  type        = string
+  default     = "refs/pull/*"
 }
 
 variable "wif_pool_id" {